Added reaction roles system
