+++ config.example.json
@@ -1,0 +1,25 @@
+{
+    "reactionRoles": [
+        {
+            "messageId": "MESSAGE_ID_WITH_REACTIONS",
+            "roleEmojis": [
+                {
+                    "emoji": "🌹",
+                    "roleId": "ROLE ID OF ROLE TO ASSIGN"
+                },
+                {
+                    "emoji": "🐞",
+                    "roleId": "ROLE ID OF ROLE TO ASSIGN"
+                },
+                {
+                    "emoji": "🪱",
+                    "roleId": "ROLE ID OF ROLE TO ASSIGN"
+                },
+                {
+                    "emoji": "🍄",
+                    "roleId": "ROLE ID OF ROLE TO ASSIGN"
+                }
+            ]
+        }
+    ]
+}
