+++ config.json
@@ -1,0 +1,193 @@
+{
+    "reactionRoles": [
+        {
+            "messageId": "1305286955050733618",
+            "roleEmojis": [
+                {
+                    "emoji": "🌹",
+                    "roleId": "1305287338850779307"
+                },
+                {
+                    "emoji": "🐞",
+                    "roleId": "1305287437651542046"
+                },
+                {
+                    "emoji": "🪱",
+                    "roleId": "1305287385063620640"
+                },
+                {
+                    "emoji": "🍄",
+                    "roleId": "1305287299285651557"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828132786012297",
+            "roleEmojis": [
+                {
+                    "emoji": "🐻",
+                    "roleId": "1269796097379930284"
+                },
+                {
+                    "emoji": "🔥",
+                    "roleId": "1269796199230214215"
+                },
+                {
+                    "emoji": "🌟",
+                    "roleId": "1269796327047430195"
+                },
+                {
+                    "emoji": "⭐",
+                    "roleId": "1269796380776333395"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828161680576583",
+            "roleEmojis": [
+                {
+                    "emoji": "🍋",
+                    "roleId": "1269796446631231569"
+                },
+                {
+                    "emoji": "🍌",
+                    "roleId": "1269796504097394710"
+                },
+                {
+                    "emoji": "🐣",
+                    "roleId": "1269796544651989053"
+                },
+                {
+                    "emoji": "🐥",
+                    "roleId": "1269796602583715921"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828184980066386",
+            "roleEmojis": [
+                {
+                    "emoji": "🥦",
+                    "roleId": "1269797052120961044"
+                },
+                {
+                    "emoji": "🍀",
+                    "roleId": "1269797104751087646"
+                },
+                {
+                    "emoji": "🍃",
+                    "roleId": "1269797154248069193"
+                },
+                {
+                    "emoji": "🦖",
+                    "roleId": "1269797225731457158"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828214314893313",
+            "roleEmojis": [
+                {
+                    "emoji": "🪼",
+                    "roleId": "1269797847809921065"
+                },
+                {
+                    "emoji": "🐳",
+                    "roleId": "1269798059550703616"
+                },
+                {
+                    "emoji": "🐋",
+                    "roleId": "1269798121399914607"
+                },
+                {
+                    "emoji": "🐟",
+                    "roleId": "1269798270805217311"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828243674894448",
+            "roleEmojis": [
+                {
+                    "emoji": "🫐",
+                    "roleId": "1269798756015149217"
+                },
+                {
+                    "emoji": "🌊",
+                    "roleId": "1269798797714919496"
+                },
+                {
+                    "emoji": "💧",
+                    "roleId": "1269798851699802174"
+                },
+                {
+                    "emoji": "🫧",
+                    "roleId": "1269798892099207198"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828266433183764",
+            "roleEmojis": [
+                {
+                    "emoji": "🍇",
+                    "roleId": "1269798947560357988"
+                },
+                {
+                    "emoji": "🪻",
+                    "roleId": "1269799061939028110"
+                },
+                {
+                    "emoji": "🐙",
+                    "roleId": "1269799103143870594"
+                },
+                {
+                    "emoji": "🐛",
+                    "roleId": "1269799235885334599"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828287832526870",
+            "roleEmojis": [
+                {
+                    "emoji": "🦑",
+                    "roleId": "1269800066349011004"
+                },
+                {
+                    "emoji": "🪷",
+                    "roleId": "1269800117930819618"
+                },
+                {
+                    "emoji": "🌺",
+                    "roleId": "1269800163858317342"
+                },
+                {
+                    "emoji": "🌷",
+                    "roleId": "1269800223496994948"
+                }
+            ]
+        },
+        {
+            "messageId": "1269828161680576583",
+            "roleEmojis": [
+                {
+                    "emoji": "🐈‍⬛",
+                    "roleId": "1269828419747844259"
+                },
+                {
+                    "emoji": "🐨",
+                    "roleId": "1269828467781013534"
+                },
+                {
+                    "emoji": "🐇",
+                    "roleId": "1269828489679212607"
+                },
+                {
+                    "emoji": "☁️",
+                    "roleId": "1269828515621109801"
+                }
+            ]
+        }
+    ]
+}
